college project
Bed Management in this pandemic has become an issue that almost all of the families of covid patients face. But by using the latest technology this problem can be solved, the main reason for this problem is hospitals not having a proper portal for managing beds. Our solution to this problem can be used to save many lives of covid patients. Therefore, we have tried to address this problem with help of our project where the focus is to make sure that every covid patient gets a bed.